+++ packages/skeleton-react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @skeletonlabs/skeleton-react
 
+## 4.1.0
+
+### Minor Changes
+
+- feat: toggle group ([#3929](https://github.com/skeletonlabs/skeleton/pull/3929))
+
+### Patch Changes
+
+- Updated dependencies [[`f5e61dd`](https://github.com/skeletonlabs/skeleton/commit/f5e61dd59aa9a4c461c129ae026add755ba7fe74)]:
+  - @skeletonlabs/skeleton-common@4.1.0
+
 ## 4.0.2
 
 ### Patch Changes