+++ packages/skeleton/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @skeletonlabs/skeleton
 
+## 4.1.0
+
 ## 4.0.2
 
 ## 4.0.1